refactor: useGetProducts에서 pageParam 사용 제거
- 사용되지 않는 pageParam 변수 제거
- useGetProducts 훅에서 queryFn 수정
- ESLint 오류 해결
